[ENTSWM-201] don't use @CreateSwarm (#24)
* [ENTSWM-201] don't use @CreateSwarm

The `@CreateSwarm` annotation is deprecated and we shouldn't
use it in boosters tests. Fortunately, its usage was limited
to only enabling a correct profile for unit tests, which can
be achieved using a system property.

* skip Surefire in openshift-it profile as well

Surefire is skipped in the `openshift` profile already, but when
running only in `openshift-it`, unit tests would be executed
and would obviously fail. This commit skips Surefire in the
`openshift-it` profile as well.